+++ Product.h
@@ -1,29 +1,30 @@
 //
-// Created by maciek on 16.12.2020.
+// Created by Maciej Bąk on 18.11.2020.
+// Version 2.0 - updated: 22.12.2020
 //
 
-#ifndef SHOPPING_LIST_PRODUCT_H
-#define SHOPPING_LIST_PRODUCT_H
+#ifndef PRODUCT_H
+#define PRODUCT_H
+
 #include <string>
+#include <utility>
 
 class Product {
 protected:
     std::string _name;
-    //static const std::string _unit;
     std::string _unit;
     int _quantity;
 public:
-    explicit Product(std::string name = "unknown", int quantity = 0) : _name(name), _quantity(quantity), _unit("undefined") {}
+    explicit Product(std::string name = "unknown", int quantity = 0) : _name(std::move(name)), _quantity(quantity), _unit("undefined") {}
     virtual ~Product()= default;
     std::string getName();
     std::string getUnit();
-    int getQuantity();
+    int getQuantity() const;
     void setQuantity(int quantity);
     void setName(std:: string& name);
-    virtual void printDetails() = 0;
 
-    virtual Product* createNew(std::string name, int quantity) = 0;
-
+    virtual void printDetails() = 0; // for printing
+    virtual Product* createNew(std::string name, int quantity) = 0; // to create new Product
 };
 
-#endif //SHOPPING_LIST_PRODUCT_H
+#endif // PRODUCT_H
