+++ theShoppingList.h
@@ -1,32 +1,34 @@
+////
+//// Created by maciek on 19.12.2020.
+////
 //
-// Created by maciek on 19.12.2020.
+//#ifndef SHOPPING_LIST_THESHOPPINGLIST_H
+//#define SHOPPING_LIST_THESHOPPINGLIST_H
+//#include <iostream>
+//#include <string>
+//#include <vector>
+//#include <iomanip> // setw, setfill
+//#include <limits> // cin.ignore argument
+//#include <climits> // INT_MAX
+//#include "ShoppingList.h"
+//#include "theShoppingList.h"
 //
-
-#ifndef SHOPPING_LIST_THESHOPPINGLIST_H
-#define SHOPPING_LIST_THESHOPPINGLIST_H
-#include <iostream>
-#include <string>
-#include <vector>
-#include <iomanip> // setw, setfill
-#include <limits> // cin.ignore argument
-#include <climits> // INT_MAX
-#include "ShoppingList.h"
-#include "theShoppingList.h"
-
-// input functions
-std::string getStringLine ();
-int getInt ();
-//int getAndStoreFloat();
-// user interaction functions
-void addShoppingList(std::vector<ShoppingList>& list);
-void deleteShoppingList(std::vector<ShoppingList>& list);
-void editShoppingList(std::vector<ShoppingList>& list);
-void showShoppingList(std::vector<ShoppingList>& list);
-void showShoppingLists(std::vector<ShoppingList>& list);
-void printMainMenu();
-void printSubMenu();
-
-void addProductToShoppingList(std::vector<ShoppingList>& list, int listPos);
-void mergeShoppingLists(std::vector<ShoppingList>& list, int listToMerge);
-void moveProductToAnotherShoppingList(std::vector<ShoppingList>& lists, int fromIndex);
-#endif //SHOPPING_LIST_THESHOPPINGLIST_H
+//// input functions
+//std::string getStringLine (bool isFile);
+//int getInt (bool isFile);
+//int getAndStoreFloat(bool isFile);
+////int getAndStoreFloat();
+//// user interaction functions
+//void addShoppingList(std::vector<ShoppingList>& list);
+//void deleteShoppingList(std::vector<ShoppingList>& list);
+//void editShoppingList(std::vector<ShoppingList>& list);
+//void showShoppingList(std::vector<ShoppingList>& list);
+//void showShoppingLists(std::vector<ShoppingList>& list);
+//void printMainMenu();
+//void printSubMenu();
+//
+//void addProductToShoppingList(std::vector<ShoppingList>& list, int listPos);
+//void mergeShoppingLists(std::vector<ShoppingList>& list, int listToMerge);
+//void moveProductToAnotherShoppingList(std::vector<ShoppingList>& lists, int fromIndex);
+//
+//#endif //SHOPPING_LIST_THESHOPPINGLIST_H
